+++ fbreader/src/formats/html/HtmlDescriptionReader.h
@@ -34,7 +34,7 @@ protected:
 	void endDocumentHandler();
 
 	bool tagHandler(const HtmlTag &tag);
-	bool characterDataHandler(const char *text, int len, bool convert);
+	bool characterDataHandler(const char *text, size_t len, bool convert);
 
 private:
 	bool myReadTitle;
