+++ fbreader/src/formats/oeb/NCXReader.h
@@ -47,7 +47,7 @@ public:
 private:
 	void startElementHandler(const char *tag, const char **attributes);
 	void endElementHandler(const char *tag);
-	void characterDataHandler(const char *text, int len);
+	void characterDataHandler(const char *text, size_t len);
 
 private:
 	BookReader &myModelReader;
