+++ fbreader/src/migration/FB2MigrationReader.cpp
@@ -28,7 +28,7 @@
 FB2MigrationReader::FB2MigrationReader(BookInfo &info, bool updateSeries) : myInfo(info), myUpdateSeries(updateSeries), myUpdateTags(info.TagsOption.value().empty()) {
 }
 
-void FB2MigrationReader::characterDataHandler(const char *text, int len) {
+void FB2MigrationReader::characterDataHandler(const char *text, size_t len) {
 	if (myReadState == READ_GENRE) {
 		myGenreBuffer.append(text, len);
 	}
